refactor utils and types
