+++ apps/photos/src/services/searchService.ts
@@ -24,12 +24,8 @@ import { Person, Thing } from 'types/machineLearning';
 import { getUniqueFiles } from 'utils/file';
 import { User } from 'types/user';
 import { getData, LS_KEYS } from 'utils/storage/localStorage';
-import {
-    EntityType,
-    LocationTag,
-    LocationTagData,
-    getLatestEntities,
-} from './entityService';
+import { getLatestEntities } from './entityService';
+import { LocationTag, LocationTagData, EntityType } from 'types/entity';
 
 const DIGITS = new Set(['0', '1', '2', '3', '4', '5', '6', '7', '8', '9']);
